Refactor: Update namespaces
