+++ src/Concerns/HasDynamicTitle.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Hyde\Framework\Models;
+namespace Hyde\Framework\Concerns;
 
 use Hyde\Framework\Hyde;
+use function Hyde\Framework\Models\str_starts_with;
 
 /**
  * Find and return the title to use for a Markdown Document.
